+++ src/App.js
@@ -1,88 +1,44 @@
-import { useState } from "react";
+import React from "react";
+import { BrowserRouter, NavLink, Route, Switch } from "react-router-dom";
+
+import SpinButton from "./components/SpinButton";
+import Navigation from "./components/Navigation";
+import Carousel from "./components/Carousel";
 
 const App = () => {
-  const PASSENGER_MAX_COUNT = 3;
-  const PASSENGER_MIN_COUNT = 1;
-
-  const [adultPassengerCount, setAdultPassengerCount] = useState(
-    PASSENGER_MIN_COUNT,
-  );
-
-  const onDecreaseCount = () => {
-    if (adultPassengerCount < PASSENGER_MIN_COUNT) return;
-
-    setAdultPassengerCount((count) => count - 1);
-  };
-
-  const onIncreaseCount = () => {
-    if (adultPassengerCount >= PASSENGER_MAX_COUNT) return;
-
-    setAdultPassengerCount((count) => count + 1);
-  };
-
   return (
-    <section>
-      <h1>승객 선택</h1>
+    <BrowserRouter>
+      <nav>
+        <ul>
+          <li>
+            <NavLink to="/spinButton">스핀버튼</NavLink>
+          </li>
+          <li>
+            <NavLink to="/carousel">캐러셀</NavLink>
+          </li>
+          <li>
+            <NavLink to="/navigation">네비게이션</NavLink>
+          </li>
+        </ul>
+      </nav>
 
-      <h2>성인</h2>
-      <label>
-        <input
-          type="button"
-          aria-label="성인 탑승자 한명 줄이기"
-          value="-"
-          disabled={adultPassengerCount < PASSENGER_MIN_COUNT}
-          onClick={onDecreaseCount}
-          style={{
-            fontSize: "2rem",
-            padding: "0 1rem",
-            margin: "0.5rem 0",
-            boxSizing: "border-box",
-          }}
-        />
+      <main>
+        <h1>접근성을 준수한 컴포넌트입니다.</h1>
+        <p>스크린리더를 활용하여 사용해보세요</p>
 
-        <span aria-atomic="true" aria-live="assertive">
-          <span
-            style={{
-              display: "inline-block",
-              width: "1px",
-              height: "1px",
-              margin: "-1px 0 0",
-              overflow: "hidden",
-            }}
-          >
-            성인 승객 {adultPassengerCount} 명
-          </span>
-          <input
-            style={{
-              fontSize: "2rem",
-              padding: "0 1rem",
-              margin: "0.5rem 0",
-              boxSizing: "border-box",
-            }}
-            value={adultPassengerCount}
-            onChange={({ target: { value } }) => setAdultPassengerCount(value)}
-            type="number"
-            min={PASSENGER_MIN_COUNT}
-            max={PASSENGER_MAX_COUNT}
-            readOnly
-          />
-        </span>
-
-        <input
-          type="button"
-          aria-label="성인 탑승자 한명 늘리기"
-          value="+"
-          disabled={adultPassengerCount >= PASSENGER_MAX_COUNT}
-          onClick={onIncreaseCount}
-          style={{
-            fontSize: "2rem",
-            padding: "0 1rem",
-            margin: "0.5rem 0",
-            boxSizing: "border-box",
-          }}
-        />
-      </label>
-    </section>
+        <Switch>
+          <Route exact path="/spinButton">
+            <SpinButton />
+          </Route>
+          <Route exact path="/carousel">
+            <Carousel />
+          </Route>
+          <Route exact path="/navigation">
+            <Navigation />
+          </Route>
+        </Switch>
+      </main>
+    </BrowserRouter>
   );
 };
