+++ src/components/Carousel.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Carousel = () => {
+  return <div>캐러셀 페이지입니다.</div>;
+};
+
+export default Carousel;
