+++ src/components/Navigation.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Navigation = () => {
+  return <div>네비게이션 영역입니다.</div>;
+};
+
+export default Navigation;
